修改后端接口错误处理

+++ src/backend-logi/defake-logi-backend/src/main/java/com/group6/defakelogibackend/service/impl/NotificationServiceImpl.java
@@ -1,6 +1,9 @@
 package com.group6.defakelogibackend.service.impl;
 
+import com.group6.defakelogibackend.exception.EntityMissingException;
+import com.group6.defakelogibackend.exception.FieldMissingException;
 import com.group6.defakelogibackend.mapper.NotificationMapper;
+import com.group6.defakelogibackend.mapper.UserMapper;
 import com.group6.defakelogibackend.model.Notification;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,21 +16,26 @@ import java.util.List;
 public class NotificationServiceImpl implements com.group6.defakelogibackend.service.NotificationService {
     @Autowired
     NotificationMapper notificationMapper;
+    @Autowired
+    UserMapper userMapper;
 
     @Override
     @Transactional
     public boolean deleteNotification(long notificationId) {
         // 检查是否存在该ID的消息
-        if (notificationMapper.findNotificationById(notificationId) != null) {
-            notificationMapper.deleteNotificationById(notificationId);
-            return true;
+        if (notificationMapper.findNotificationById(notificationId) == null) {
+            throw new EntityMissingException("notificationId 错误，不存在对应的notification!");
         }
-        return false;
+        notificationMapper.deleteNotificationById(notificationId);
+        return true;
     }
 
     @Override
     @Transactional
     public boolean deleteAllNotification(long userId) {
+        if (userMapper.findUserById(userId) == null){
+            throw new EntityMissingException("userId 错误，不存在对应的user!");
+        }
         notificationMapper.deleteAllNotificationByUserId(userId);
         return true;
     }
@@ -35,18 +43,28 @@ public class NotificationServiceImpl implements com.group6.defakelogibackend.ser
     @Override
     @Transactional
     public Notification notificationInfo(long notificationId) {
-        return notificationMapper.findNotificationById(notificationId);
+        Notification notification = notificationMapper.findNotificationById(notificationId);
+        if (notification == null) {
+            throw new EntityMissingException("notificationId 错误, 找不到对应的通知!");
+        }
+        return notification;
     }
 
     @Override
     @Transactional
     public List<Notification> notificationFilter(long userId, String condition) {
-        if (condition.equals("send_at_desc")) {
+        if (userMapper.findUserById(userId) == null){
+            throw new EntityMissingException("userId 错误，不存在对应的user!");
+        }
+        System.out.println(condition);
+        System.out.println();
+        System.out.println();
+        if (condition.equals("sent_at_desc")) {
             return notificationMapper.findNotificationByUserId_sent_at_desc(userId);
-        } else if (condition.equals("send_at_asc")) {
+        } else if (condition.equals("sent_at_asc")) {
             return notificationMapper.findNotificationByUserId_sent_at_asc(userId);
         } else {
-            return null;
+            throw new FieldMissingException("输入的 condition 字段暂不支持!");
         }
     }
 }
